Created Delete Article Screen. Added more article functions.
Added Delete, Backup, and RestoreBackup (remove the article table and
restore the backup) functions.

Delete Article screen now operational.

Added setSessionRole and getSessionRole functions in the HelpSystem java
file. Example: If the user is logged in as an admin, then their session
role will be "admin" and so on.

Added getter and setter functions in User java file to implement
sessionRole functions in HelpSystem.java